* origin/main:
  Optimize Docker build with dependency layer caching (#36864)
  Fix URLJoin, markup render link reoslving, sign-in/up/linkaccount page common data (#36861)
  Fix CodeQL code scanning alerts (#36858)
  Refactor auth middleware (#36848)
  Update Nix flake (#36857)
  Update JS deps (#36850)
  Load `mentionValues` asynchronously (#36739)
  [skip ci] Updated translations via Crowdin

+++ modules/indexer/code/search.go
@@ -130,7 +130,7 @@ func searchResult(result *internal.SearchResult, startIndex, endIndex int) (*Res
 }
 
 // PerformSearch perform a search on a repository
-func PerformSearch(ctx context.Context, opts *SearchOptions) (int, []*Result, []*SearchResultLanguages, error) {
+func PerformSearch(ctx context.Context, opts *SearchOptions) (int64, []*Result, []*SearchResultLanguages, error) {
 	if opts == nil || len(opts.Keyword) == 0 {
 		return 0, nil, nil, nil
 	}
@@ -149,5 +149,5 @@ func PerformSearch(ctx context.Context, opts *SearchOptions) (int, []*Result, []
 			return 0, nil, nil, err
 		}
 	}
-	return int(total), displayResults, resultLanguages, nil
+	return total, displayResults, resultLanguages, nil
 }
